server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

faster-whisper==0.10.0

torch==2.1.0
torchaudio==2.1.0

numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

pyannote.audio==3.1.1

nltk==3.8.1
sumy==0.11.0

pymongo==4.6.0
python-multipart==0.0.6
